Logical sharding for input tensor and halo output
To reach parity on pytorch traces between input
tensor being on device vs being on host, logical
sharding is forced in case input tensor is interleaved
and I2S needs to be run.
If tensor is on host then ttnn::to_device is used
and tesor being allocated in that codepath is going
to use way less memory (x2) in many cases.
To reach parithy with to_device codepath, logical
sharding is used for I2S codepath.

Output of halo op is also switched to use logical
sharding which saves memory in many cases.

Output of the conv2d op is forced to be physical
sharding to avoid failiours in models as subsequent
op after conv2d won't handle logical sharding properly
in many cases.

Switching to logical sharding in output would save
even more memory in some case, and should eventually
be done.

Before this change number of failing cases in pytorch
traces if input is on device was 26 and with this change
is 16 same as if input is on host.